move cylc helper scripts in correct place
