+++ wrfpy/cylc/wps_post.py
@@ -4,7 +4,7 @@ import argparse
 import datetime
 import time
 from wrfpy import utils
-from config import config
+from wrfpy.config import config
 import os
 #from urb import urb
 import shutil
@@ -23,12 +23,12 @@ class wps_post(config):
         ## wrf run dir
         # cleanup old met_em files
         # create list of files to remove
-        files = [glob.glob(os.path.join(rundir, ext))
-                 for ext in ['met_em*']]
+        #files = [glob.glob(os.path.join(rundir, ext))
+        #         for ext in ['met_em*']]
         # flatten list
-        files_flat = [item for sublist in files for item in sublist] 
+        #files_flat = [item for sublist in files for item in sublist] 
         # remove files silently
-        [ utils.silentremove(filename) for filename in files_flat ]
+        #[ utils.silentremove(filename) for filename in files_flat ]
         # copy new met_em files
         # create list of files to copy
         files = [glob.glob(os.path.join(wpsdir, ext))
@@ -44,8 +44,6 @@ class wps_post(config):
         files_flat = [item for sublist in files for item in sublist]
         # remove files silently
         [ utils.silentremove(filename) for filename in files_flat ]
-
-
 
 if __name__=="__main__":
     wps_post()
