--- wrfpy/cylc/wps_ungrib.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-
-import argparse
-import datetime
-import time
-
-time.sleep(1)
